refactor(cli): extract intake commands into cli_capture.py (2b)
mine-history / ingest / capture-stop / resume (+ their ingest helpers
_resolve_ingest_row/_is_high_signal/_extract_first_h1 and the high-signal
tag/URL constants, none used elsewhere) moved out of cli.py. cli.py 2100 →
1291 LOC. ruff + mypy clean.
